dynamic import of rigidbody
