+++ apps/web/src/components/ThreeHero/index.tsx
@@ -10,6 +10,19 @@ import dynamic from 'next/dynamic';
 import Image, { StaticImageData } from 'next/image';
 import Link from 'apps/web/src/components/Link';
 
+// 3D libraries - types
+import type { Vector3 } from '@react-three/fiber';
+import type { Vector3Tuple } from '@react-three/rapier';
+import type {
+  ColorRepresentation,
+  Mesh,
+  BufferGeometry,
+  NormalBufferAttributes,
+  Material,
+} from 'three';
+
+import { Bloom, SMAA, EffectComposer } from '@react-three/postprocessing';
+
 // Assets
 import {
   BaseLogo,
@@ -30,17 +43,6 @@ import {
 
 import baseLogo from './assets/base-logo.svg';
 import environmentLight from './assets/environmentLight.jpg';
-
-// 3D libraries - types
-import type { Vector3 } from '@react-three/fiber';
-import type { Vector3Tuple } from '@react-three/rapier';
-import type {
-  ColorRepresentation,
-  Mesh,
-  BufferGeometry,
-  NormalBufferAttributes,
-  Material,
-} from 'three';
 
 // 3D libraries - dynamic imports
 
@@ -70,20 +72,6 @@ const Environment = dynamic(
 
 // Dynamic - react-three/rapier
 const Physics = dynamic(async () => import('@react-three/rapier').then((mod) => mod.Physics), {
-  ssr: false,
-});
-
-// Dynamic - react-three/postprocessing
-const EffectComposer = dynamic(
-  async () => import('@react-three/postprocessing').then((mod) => mod.EffectComposer),
-  { ssr: false },
-);
-
-const Bloom = dynamic(async () => import('@react-three/postprocessing').then((mod) => mod.Bloom), {
-  ssr: false,
-});
-
-const SMAA = dynamic(async () => import('@react-three/postprocessing').then((mod) => mod.SMAA), {
   ssr: false,
 });
